requirements.txt
# Python version requirement: 3.8+ (remove enum34 as obsolete in modern Python)
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==15.0.1
python-multipart==0.0.6
httpx==0.28.1

# Google ADK and Vertex AI
google-adk @ git+https://github.com/google/adk-python.git@main
google-cloud-aiplatform==1.38.0
google-auth==2.25.2

# Multi-agent system dependencies
# NOTE: asyncio, dataclasses, and enum are built into modern Python (3.7+/3.4+)

# Supabase integration
supabase==2.17.0
gotrue==2.12.3
postgrest==1.1.1
storage3==0.12.0
supafunc==0.10.1
realtime==2.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
# NOTE: httpx already specified above - using 0.28.1 for consistency

# RAG and clustering dependencies
scikit-learn==1.3.2  # Consider 1.4+ for latest features
numpy==1.24.3  # Consider 1.26+ for Python 3.12+ compatibility
sentence-transformers==2.2.2

# Development
python-dotenv==1.0.0

# Observability
opentelemetry-exporter-otlp==1.35.0